benchmarks/requirements.txt
mem0ai>=1.0.5
opensearch-py>=2.4.0
psycopg2-binary>=2.9.9
pytest>=8.0.0
httpx>=0.27.0
